1. Add 'listPengumpulanByNama' at PengumpulanDataController 2. Fixing 'VerifySipastiJwt' type for 'superadmn' to 'superadmin' 3. Fixing minor bug for the migration

+++ database/migrations/2025_08_19_035615_create_sumber_daya_vendor_table.php
@@ -13,16 +13,18 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('sumber_daya_vendor', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('data_vendor_id')->nullable();
-            $table->string('jenis')->nullable();
-            $table->string('nama')->nullable();
-            $table->string('spesifikasi')->nullable();
-            $table->timestamps();
+        if(!Schema::hasTable('sumber_daya_vendor')){
+            Schema::create('sumber_daya_vendor', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('data_vendor_id')->nullable();
+                $table->string('jenis')->nullable();
+                $table->string('nama')->nullable();
+                $table->string('spesifikasi')->nullable();
+                $table->timestamps();
 
-            $table->foreign('data_vendor_id')->references('id')->on('data_vendors')->onDelete('set null');
-        });
+                $table->foreign('data_vendor_id')->references('id')->on('data_vendors')->onDelete('set null');
+            });
+        }
     }
 
     /**
